+++ backend/insurance/src/test/java/com/innov8ors/insurance/util/TestUtil.java
@@ -37,14 +37,14 @@ public class TestUtil {
     public static final String TEST_POLICY_NAME = "Test Policy";
     public static final String TEST_POLICY_TYPE = "Health";
     public static final String TEST_POLICY_DESCRIPTION = "This is a test policy description.";
-    public static final Double TEST_POLICY_PREMIUM_AMOUNT = 100.0;
-    public static final Double TEST_POLICY_COVERAGE_AMOUNT = 10000.0;
+    public static final BigDecimal TEST_POLICY_PREMIUM_AMOUNT = BigDecimal.valueOf(100);
+    public static final BigDecimal TEST_POLICY_COVERAGE_AMOUNT = BigDecimal.valueOf(10000);
     public static final Integer TEST_POLICY_DURATION_MONTHS = 12; // Assuming a default duration of 12 months
     public static final Long TEST_USER_POLICY_ID = 2L;
     public static final LocalDateTime TEST_POLICY_START_DATE = LocalDateTime.now();
     public static final LocalDateTime TEST_POLICY_END_DATE = LocalDateTime.now().plusMonths(TEST_POLICY_DURATION_MONTHS);
     public static final UserPolicyStatus TEST_USER_POLICY_STATUS = UserPolicyStatus.ACTIVE;
-    public static final BigDecimal TEST_POLICY_PREMIUM_PAID = BigDecimal.valueOf(TEST_POLICY_PREMIUM_AMOUNT);
+    public static final BigDecimal TEST_POLICY_PREMIUM_PAID = TEST_POLICY_PREMIUM_AMOUNT;
 
 
     public static User getUser() {
